fix: extract HTML template into typed data + renderer
- Creates src/data.ts with TypeScript interfaces (Project, TimelineEntry,
  WalletInfo, Service, Collaborator) and typed arrays for all site content
- Creates src/render.ts with renderProjectCard, renderServiceCard,
  renderCollaboratorCard, renderTimeline, renderWalletCard, and renderHTML
- Updates src/index.ts to import from data.ts and render.ts instead of
  embedding 200+ lines of HTML/CSS inline

Closes #5
