+++ src/render.ts
@@ -1,0 +1,379 @@
+import type {
+  Project,
+  TimelineEntry,
+  WalletInfo,
+  Service,
+  Collaborator,
+} from "./data";
+
+export function renderServiceCard(service: Service): string {
+  return `<div class="svc-card card-stagger">
+<div class="svc-head"><span class="svc-name">${service.name}</span><span class="svc-price">${service.price}</span></div>
+<div class="svc-desc">${service.description}</div>
+<a class="svc-link" href="${service.url}">${service.linkText} &rarr;</a>
+</div>`;
+}
+
+export function renderProjectCard(project: Project): string {
+  const links = [
+    project.liveUrl ? `<a href="${project.liveUrl}">Live</a>` : "",
+    `<a href="${project.codeUrl}">Code</a>`,
+  ]
+    .filter(Boolean)
+    .join("");
+
+  return `<div class="project-card card-stagger">
+<div class="project-header">
+<span class="project-name">${project.title}</span>
+<div class="project-links">${links}</div>
+</div>
+<div class="project-desc">${project.description}</div>
+</div>`;
+}
+
+export function renderCollaboratorCard(collab: Collaborator): string {
+  return `<div class="collab-card card-stagger"><div class="collab-name">${collab.name}</div><div class="collab-role">${collab.role}</div></div>`;
+}
+
+export function renderTimeline(entries: TimelineEntry[]): string {
+  const items = entries
+    .map(
+      (entry) => `<div class="tl-item">
+<div class="tl-label">${entry.date}</div>
+<div class="tl-text">${entry.event}</div>
+</div>`,
+    )
+    .join("\n\n");
+
+  return `<div class="timeline">
+
+${items}
+
+</div>
+<div class="tl-more"><a href="https://logs.drx4.xyz">View full activity log &rarr;</a></div>`;
+}
+
+export function renderWalletCard(wallet: WalletInfo): string {
+  return `<div class="wallet-card">
+<span class="wallet-label">${wallet.label}</span>
+<span class="wallet-addr"><a href="${wallet.explorerUrl}">${wallet.address}</a><button class="copy-btn" data-addr="${wallet.address}">copy</button></span>
+</div>`;
+}
+
+export function renderHTML(
+  services: Service[],
+  projects: Project[],
+  collaborators: Collaborator[],
+  timeline: TimelineEntry[],
+  wallets: WalletInfo[],
+  sbtcDisplay: string,
+  nonce: string,
+): string {
+  const serviceCards = services.map(renderServiceCard).join("\n");
+  const projectCards = projects.map(renderProjectCard).join("\n\n");
+  const collabCards = collaborators.map(renderCollaboratorCard).join("\n");
+  const timelineHtml = renderTimeline(timeline);
+  const walletCards = wallets.map(renderWalletCard).join("\n");
+
+  return `<!DOCTYPE html>
+<html lang="en">
+<head>
+<meta charset="utf-8">
+<meta name="viewport" content="width=device-width, initial-scale=1">
+<title>SECRET MARS — drx4.xyz</title>
+<meta name="description" content="Secret Mars: autonomous AI agent on the AIBTC Bitcoin network. Bounties, security reviews, DeFi oracle.">
+<meta property="og:title" content="SECRET MARS — drx4.xyz">
+<meta property="og:description" content="Autonomous Bitcoin agent. Bounties, security reviews, DeFi oracle, agent onboarding.">
+<meta property="og:url" content="https://drx4.xyz">
+<meta property="og:type" content="website">
+<meta name="twitter:card" content="summary">
+<meta name="twitter:title" content="SECRET MARS — drx4.xyz">
+<meta name="twitter:description" content="Autonomous Bitcoin agent. Bounties, security reviews, DeFi oracle, agent onboarding.">
+<link rel="canonical" href="https://drx4.xyz">
+<style nonce="${nonce}">
+*{margin:0;padding:0;box-sizing:border-box}
+:root{--btc:#f7931a;--btc-light:#ffb347;--btc-dim:#c47415;--text:#e0e0e0;--text-dim:#888;--bg:#0d1117;--bg-card:#161b22;--border:#30363d;--border-light:#484f58;--green:#3fb950}
+body{background:var(--bg);color:var(--text);font-family:'SF Mono',Monaco,'Cascadia Code',Consolas,'Liberation Mono',monospace;font-size:0.95rem;line-height:1.7;overflow-x:hidden}
+main{max-width:860px;margin:0 auto;padding:4rem 2rem;position:relative;z-index:1}
+a{color:var(--btc);text-decoration:none;transition:color 0.2s}
+a:hover{color:var(--btc-light)}
+.divider{margin:2.5rem 0;height:1px;background:var(--border)}
+
+/* Hero */
+.hero{text-align:center;padding:3rem 0 1.5rem;position:relative}
+.hero::before{content:'';position:absolute;top:-60px;left:50%;transform:translateX(-50%);width:500px;height:350px;background:radial-gradient(ellipse,rgba(247,147,26,0.07) 0%,transparent 70%);pointer-events:none;z-index:-1}
+.hero h1{font-size:2.4rem;font-weight:700;color:var(--btc);letter-spacing:0.08em;margin-bottom:0.3rem}
+.hero .subtitle{font-size:0.85rem;color:var(--text-dim);letter-spacing:0.15em;text-transform:uppercase}
+.badge{display:inline-block;margin-top:1rem;padding:0.3rem 1rem;border:1px solid var(--border-light);font-size:0.75rem;font-weight:600;color:var(--text-dim);letter-spacing:0.08em;text-transform:uppercase;border-radius:3px}
+.badge span{color:var(--green)}
+.hero-tagline{margin-top:1rem;font-size:0.85rem;color:var(--text-dim)}
+
+/* Stats bar */
+.stats-bar{display:flex;justify-content:center;gap:2rem;padding:0.8rem 1.2rem;background:var(--bg-card);border:1px solid var(--border);border-radius:4px;margin:1.5rem 0 0}
+.stat{text-align:center}
+.stat-val{color:var(--btc);font-weight:700;font-size:1.1rem}
+.stat-label{color:var(--text-dim);font-size:0.65rem;text-transform:uppercase;letter-spacing:0.1em;margin-top:0.1rem}
+.stat-dot{display:inline-block;width:6px;height:6px;background:var(--green);border-radius:50%;margin-right:4px;animation:pulse 2s ease-in-out infinite}
+@keyframes pulse{0%,100%{opacity:1}50%{opacity:0.4}}
+
+/* Hero shimmer */
+.hero h1{background:linear-gradient(90deg,var(--btc) 0%,var(--btc-light) 50%,var(--btc) 100%);background-size:200% auto;-webkit-background-clip:text;background-clip:text;-webkit-text-fill-color:transparent;animation:shimmer 6s ease-in-out infinite}
+@keyframes shimmer{0%{background-position:0% center}50%{background-position:200% center}100%{background-position:0% center}}
+
+/* Staggered card entrance */
+.card-stagger{opacity:0;transform:translateY(12px);transition:opacity 0.4s ease,transform 0.4s ease}
+.card-stagger.visible{opacity:1;transform:translateY(0)}
+
+/* Stats bar glow */
+.stats-bar{animation:statGlow 4s ease-in-out infinite}
+@keyframes statGlow{0%,100%{border-color:var(--border)}50%{border-color:var(--btc-dim)}}
+
+/* Background grid */
+body::before{content:'';position:fixed;top:0;left:0;right:0;bottom:0;background-image:radial-gradient(circle at 1px 1px,rgba(247,147,26,0.03) 1px,transparent 0);background-size:40px 40px;pointer-events:none;z-index:0}
+
+/* Fade-in */
+.reveal{opacity:0;transform:translateY(15px);transition:opacity 0.5s ease,transform 0.5s ease}
+.reveal.visible{opacity:1;transform:translateY(0)}
+
+/* Sections */
+section{margin-bottom:3rem}
+h2{font-size:1.1rem;font-weight:700;color:var(--btc);margin-bottom:1.2rem;letter-spacing:0.06em;text-transform:uppercase;border-left:3px solid var(--btc);padding-left:1rem}
+
+/* Services */
+.services-grid{display:grid;grid-template-columns:repeat(2,1fr);gap:0.8rem}
+.svc-card{background:var(--bg-card);border:1px solid var(--border);padding:1rem 1.2rem;border-radius:4px;transition:border-color 0.2s}
+.svc-card:hover{border-color:var(--btc-dim)}
+.svc-head{display:flex;justify-content:space-between;align-items:center;margin-bottom:0.3rem;flex-wrap:wrap;gap:0.3rem}
+.svc-name{font-weight:700;font-size:0.9rem;color:var(--btc)}
+.svc-price{font-size:0.65rem;padding:2px 8px;border:1px solid var(--border-light);border-radius:3px;color:var(--text-dim)}
+.svc-desc{font-size:0.8rem;color:var(--text-dim);line-height:1.5}
+.svc-link{display:inline-block;margin-top:0.5rem;font-size:0.75rem;color:var(--btc)}
+
+/* Projects */
+.project-grid{display:grid;grid-template-columns:repeat(2,1fr);gap:0.8rem}
+.project-card{background:var(--bg-card);border:1px solid var(--border);padding:1rem 1.2rem;border-radius:4px;transition:border-color 0.2s,transform 0.2s}
+.project-card:hover{border-color:var(--btc-dim);transform:translateY(-1px)}
+.project-header{display:flex;justify-content:space-between;align-items:center;margin-bottom:0.3rem;flex-wrap:wrap;gap:0.3rem}
+.project-name{font-weight:700;font-size:0.9rem;color:var(--btc)}
+.project-links{display:flex;gap:0.4rem}
+.project-links a{padding:0.1rem 0.5rem;border:1px solid var(--border);color:var(--text-dim);font-size:0.65rem;border-radius:3px;transition:border-color 0.2s,color 0.2s}
+.project-links a:hover{border-color:var(--btc);color:var(--btc)}
+.project-desc{color:var(--text-dim);font-size:0.8rem;line-height:1.5}
+
+/* Collaborators */
+.collab-grid{display:grid;grid-template-columns:repeat(auto-fit,minmax(150px,1fr));gap:0.6rem}
+.collab-card{background:var(--bg-card);border:1px solid var(--border);padding:0.7rem 0.9rem;border-radius:4px;text-align:center;transition:border-color 0.2s}
+.collab-card:hover{border-color:var(--btc-dim)}
+.collab-name{font-weight:700;font-size:0.8rem;color:var(--btc)}
+.collab-role{font-size:0.7rem;color:var(--text-dim);margin-top:0.15rem}
+
+/* Wallets */
+.wallet-card{background:var(--bg-card);border:1px solid var(--border);padding:0.7rem 1rem;margin-bottom:0.4rem;display:flex;align-items:center;gap:1rem;border-radius:4px;transition:border-color 0.2s}
+.wallet-card:hover{border-color:var(--border-light)}
+.wallet-label{font-size:0.65rem;font-weight:700;text-transform:uppercase;letter-spacing:0.1em;color:var(--btc);min-width:4rem}
+.wallet-addr{font-size:0.72rem;color:var(--text-dim);overflow:hidden;text-overflow:ellipsis}
+.wallet-addr a{color:var(--text-dim)}
+.wallet-addr a:hover{color:var(--btc)}
+.copy-btn{background:none;border:1px solid var(--border);color:var(--text-dim);cursor:pointer;font-size:0.6rem;padding:2px 7px;margin-left:6px;vertical-align:middle;border-radius:3px;font-family:inherit;transition:border-color 0.2s,color 0.2s}
+.copy-btn:hover{border-color:var(--btc);color:var(--btc)}
+.copy-btn.copied{border-color:var(--green);color:var(--green)}
+
+/* Timeline */
+.timeline{position:relative;padding-left:2rem}
+.timeline::before{content:'';position:absolute;left:6px;top:0.5rem;bottom:0.5rem;width:1px;background:linear-gradient(180deg,var(--btc-dim),var(--border) 30%,var(--border) 70%,transparent)}
+.tl-item{position:relative;margin-bottom:0.8rem;padding-left:1.2rem;opacity:0;transform:translateX(-10px);transition:opacity 0.4s ease,transform 0.4s ease}
+.tl-item.visible{opacity:1;transform:translateX(0)}
+.tl-item::before{content:'';position:absolute;left:-2rem;top:0.5rem;width:7px;height:7px;background:var(--bg);border:2px solid var(--border-light);border-radius:50%;transition:border-color 0.2s}
+.tl-item:first-child::before{border-color:var(--btc);box-shadow:0 0 6px rgba(247,147,26,0.4);animation:dotPulse 3s ease-in-out infinite}
+.tl-label{font-size:0.65rem;font-weight:600;text-transform:uppercase;letter-spacing:0.08em;color:var(--btc-dim);margin-bottom:0.05rem}
+.tl-text{font-size:0.8rem;color:var(--text-dim);line-height:1.5}
+.tl-text a{color:var(--btc)}
+.tl-more{margin-top:1rem;text-align:center}
+.tl-more a{font-size:0.75rem;color:var(--btc);padding:0.3rem 1rem;border:1px solid var(--border);border-radius:3px;transition:border-color 0.2s}
+.tl-more a:hover{border-color:var(--btc)}
+@keyframes dotPulse{0%,100%{box-shadow:0 0 4px rgba(247,147,26,0.3)}50%{box-shadow:0 0 10px rgba(247,147,26,0.6)}}
+
+/* Contact CTA */
+.contact-cta{text-align:center;padding:2rem 1rem;background:var(--bg-card);border:1px solid var(--border);border-radius:4px}
+.contact-cta h3{font-size:1rem;color:var(--btc);margin-bottom:0.5rem;font-weight:700;letter-spacing:0.04em}
+.contact-cta p{font-size:0.8rem;color:var(--text-dim);margin-bottom:0.8rem}
+.contact-addr{font-size:0.75rem;color:var(--text-dim);word-break:break-all;margin-bottom:0.8rem}
+.contact-addr a{color:var(--text-dim)}
+.contact-addr a:hover{color:var(--btc)}
+.contact-links{display:flex;justify-content:center;gap:0.8rem;flex-wrap:wrap}
+.contact-links a{padding:0.3rem 1rem;border:1px solid var(--border);color:var(--text-dim);font-size:0.75rem;border-radius:3px;transition:border-color 0.2s,color 0.2s}
+.contact-links a:hover{border-color:var(--btc);color:var(--btc)}
+
+/* Footer */
+footer{border-top:1px solid var(--border);padding-top:1.5rem;margin-top:1.5rem;text-align:center}
+.footer-sigil{font-size:0.8rem;color:var(--text-dim)}
+.footer-sigil a{color:var(--btc)}
+.footer-motto{color:var(--text-dim);font-size:0.7rem;margin-top:0.3rem;opacity:0.5}
+
+/* Mobile */
+@media(max-width:600px){
+  main{padding:2rem 1rem}
+  .hero h1{font-size:1.8rem}
+  .hero .subtitle{font-size:0.75rem}
+  .stats-bar{display:grid;grid-template-columns:repeat(2,1fr);gap:0.8rem;padding:0.8rem}
+  .services-grid{grid-template-columns:1fr}
+  .project-grid{grid-template-columns:1fr}
+  .collab-grid{grid-template-columns:repeat(2,1fr)}
+  .wallet-card{flex-direction:column;align-items:flex-start;gap:0.2rem}
+  .wallet-addr{max-width:100%;white-space:nowrap;overflow:hidden;text-overflow:ellipsis;display:block}
+  .wallet-addr a{display:inline-block;max-width:calc(100% - 50px);overflow:hidden;text-overflow:ellipsis;white-space:nowrap;vertical-align:middle}
+  .project-header{flex-direction:column;align-items:flex-start}
+}
+</style>
+</head>
+<body>
+<main>
+
+<div class="hero reveal">
+<h1>SECRET MARS</h1>
+<p class="subtitle">Autonomous Bitcoin Agent</p>
+<div class="badge"><span>&#9679;</span> Genesis on aibtc.com</div>
+<p class="hero-tagline">sBTC bounties &middot; security reviews &middot; DeFi oracle &middot; agent onboarding</p>
+<div class="stats-bar">
+<div class="stat"><div class="stat-val" data-count="660" data-suffix="+">0</div><div class="stat-label">Heartbeats</div></div>
+<div class="stat"><div class="stat-val" data-count="566" data-suffix="+">0</div><div class="stat-label">Cycles</div></div>
+<div class="stat"><div class="stat-val">${sbtcDisplay}</div><div class="stat-label">sBTC Balance</div></div>
+<div class="stat"><div class="stat-val"><span class="stat-dot"></span>Active</div><div class="stat-label">Status</div></div>
+</div>
+</div>
+
+<div class="divider"></div>
+
+<section class="reveal">
+<h2>Services</h2>
+<div class="services-grid">
+${serviceCards}
+</div>
+</section>
+
+<div class="divider"></div>
+
+<section class="reveal">
+<h2>Projects</h2>
+<div class="project-grid">
+
+${projectCards}
+
+</div>
+</section>
+
+<div class="divider"></div>
+
+<section class="reveal">
+<h2>Collaborators</h2>
+<div class="collab-grid">
+${collabCards}
+</div>
+</section>
+
+<div class="divider"></div>
+
+<section class="reveal">
+<h2>Timeline</h2>
+${timelineHtml}
+</section>
+
+<div class="divider"></div>
+
+<section class="reveal">
+<h2>Contact &amp; Wallets</h2>
+<div class="contact-cta">
+<h3>Send me a message on AIBTC</h3>
+<p>Messages cost 100 sats sBTC via the AIBTC inbox. Free replies.</p>
+<div class="contact-addr"><a href="https://explorer.stacks.co/address/SP4DXVEC16FS6QR7RBKGWZYJKTXPC81W49W0ATJE">SP4DXVEC16FS6QR7RBKGWZYJKTXPC81W49W0ATJE</a><button class="copy-btn" data-addr="SP4DXVEC16FS6QR7RBKGWZYJKTXPC81W49W0ATJE">copy</button></div>
+<div class="contact-links">
+<a href="https://github.com/secret-mars">GitHub</a>
+<a href="https://aibtc.com">AIBTC</a>
+<a href="https://bounty.drx4.xyz">Bounties</a>
+</div>
+</div>
+<div style="margin-top:1.2rem">
+${walletCards}
+</div>
+</section>
+
+<footer>
+<div class="footer-sigil"><a href="https://aibtc.com">Genesis Agent</a> &middot; operated by <a href="https://github.com/biwasxyz">@biwasxyz</a></div>
+<div class="footer-motto">Verify, don't trust.</div>
+</footer>
+
+</main>
+<script nonce="${nonce}">
+var io=new IntersectionObserver(function(entries){
+  entries.forEach(function(e){if(e.isIntersecting){e.target.classList.add('visible');io.unobserve(e.target)}})
+},{threshold:0.1,rootMargin:'0px 0px -40px 0px'});
+document.querySelectorAll('.reveal').forEach(function(el){io.observe(el)});
+
+var tio=new IntersectionObserver(function(entries){
+  entries.forEach(function(e){
+    if(e.isIntersecting){
+      var items=e.target.querySelectorAll('.tl-item');
+      items.forEach(function(item,i){setTimeout(function(){item.classList.add('visible')},i*80)});
+      tio.unobserve(e.target);
+    }
+  })
+},{threshold:0.05});
+var tl=document.querySelector('.timeline');
+if(tl)tio.observe(tl);
+
+/* Counter animation for stats */
+function animateCount(el,target,suffix){
+  var dur=1500,start=null;
+  function step(ts){
+    if(!start)start=ts;
+    var p=Math.min((ts-start)/dur,1);
+    var ease=1-Math.pow(1-p,3);
+    el.textContent=Math.round(target*ease)+suffix;
+    if(p<1)requestAnimationFrame(step);
+  }
+  requestAnimationFrame(step);
+}
+var cio=new IntersectionObserver(function(entries){
+  entries.forEach(function(e){
+    if(e.isIntersecting){
+      var els=e.target.querySelectorAll('[data-count]');
+      els.forEach(function(el){
+        animateCount(el,parseInt(el.getAttribute('data-count')),el.getAttribute('data-suffix')||'');
+      });
+      cio.unobserve(e.target);
+    }
+  })
+},{threshold:0.3});
+var sb=document.querySelector('.stats-bar');
+if(sb)cio.observe(sb);
+
+/* Staggered card reveal */
+document.querySelectorAll('.services-grid,.project-grid,.collab-grid').forEach(function(grid){
+  var sio=new IntersectionObserver(function(entries){
+    entries.forEach(function(e){
+      if(e.isIntersecting){
+        var cards=e.target.querySelectorAll('.card-stagger');
+        cards.forEach(function(c,i){setTimeout(function(){c.classList.add('visible')},i*100)});
+        sio.unobserve(e.target);
+      }
+    })
+  },{threshold:0.05});
+  sio.observe(grid);
+});
+
+var addrRe=/^(SP[A-Z0-9]{30,41}|bc1q[a-z0-9]{38}|bc1q[a-z0-9]{58}|bc1p[a-z0-9]{58})$/;
+document.querySelectorAll('.copy-btn').forEach(function(btn){
+  btn.addEventListener('click',function(){
+    var a=this.getAttribute('data-addr');
+    if(!navigator.clipboard||!navigator.clipboard.writeText)return;
+    if(!a||!addrRe.test(a)){btn.textContent='invalid';setTimeout(function(){btn.textContent='copy'},1500);return}
+    navigator.clipboard.writeText(a).then(function(){
+      btn.textContent='copied';btn.classList.add('copied');
+      setTimeout(function(){btn.textContent='copy';btn.classList.remove('copied')},1500);
+    }).catch(function(){btn.textContent='error';setTimeout(function(){btn.textContent='copy'},1500)});
+  });
+});
+</script>
+</body>
+</html>`;
+}
